requirements.txt
fastapi
uvicorn
langgraph
pydantic
chromadb
langchain
langchain-openai
requests
slowapi
langchain[google-genai]
boto3
beautifulsoup4
duckduckgo-search
rank-bm25
langfuse
